+++ CHANGES.md
@@ -1,3 +1,7 @@
+### V 1.1.1 22/06/2016
+ * Improve error details adding the stack strace of the error
+ * Fix: It was impossible to send SMS when App was started at the end of Android boot phase
+
 ### V 1.1.0 29/04/2016
  * SMServer service is now started at the end of Android boot phase
  * Add SMServer version in notification bar